requirements.txt
llama-index==0.10.1
llama-index-vector-stores-postgres==0.1.3
sentence-transformers==2.2.2
psycopg2-binary==2.9.10
pgvector==0.2.5
python-dotenv==1.0.0
numpy==1.26.4
huggingface-hub==0.16.4
# torch==2.1.0
# transformers==4.35.0
openai==1.99.9
psutil==5.9.8
